openai-rag/requirements.txt
openai
python-dotenv
pandas
numpy
faiss-cpu
tiktoken
